+++ NEWS.md
@@ -21,6 +21,9 @@
   when called on a very long `wkb()` vector (#117).
 * Fixed issue with `validate_wk_wkb()` and `validate_wk_wkt()`, which failed
   for most valid objects (#119).
+* Added `wk_envelope()` and `wk_envelope_handler()` to compute feature-wise
+  bounding boxes (#120, #122).
+* Fixed headers and tests to pass on big endian systems (#105, #122).
 
 # wk 0.5.0